load/LoaderをLoaderResponseに変更、LoadPromisesを独立
