+++ packages/maker/package.json
@@ -15,6 +15,7 @@
     "@types/react-split-pane": "^0.1.67",
     "@welldone-software/why-did-you-render": "^4.1.3",
     "@wwawing/common-interface": "^3.2.11",
+    "@wwawing/loader": "^3.4.0",
     "node-sass": "^4.12.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
@@ -30,8 +31,8 @@
     "typescript-fsa": "^3.0.0",
     "typescript-fsa-reducers": "^1.2.1",
     "typescript-fsa-redux-thunk": "^2.1.9",
-    "wwamaker-saver": "^0.0.1",
-    "wwamaker-image-decorder": "^0.0.1"
+    "wwamaker-image-decorder": "^0.0.1",
+    "wwamaker-saver": "^0.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
